Add support to visualice the image in external program
Currently con the key-map Ctrl+i it open the image with an external program. By default is eog, but it could be other program configurable in the new attribute command_open_imen, if the att exist, the image is open with the tool send; this allow the flexibility to open the image in differents programs accourding to each machine.